test(gateway): move workspace conflict lookup placement tests into their own suite
